+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# v1.9.4
+## 05/09/2019
+
+1. [](#new)
+    * Added support for `field.copy-to-clipboard` on Text input fields
+1. [](#improved)
+    * Only invalidate cache on creating new/deleting page to speed up the recovery
+    * Updated language strings from https://crowdin.com/project/grav-admin
+    * Use `plugins://` stream rather than `user://plugins` [#1674](https://github.com/getgrav/grav-plugin-admin/issues/1674)
+1. [](#bugfix)
+    * Fixed admin cache to detect moved and deleted pages
+    * Fixed avatar URLs with `?` in them being broken
+    * Fixed issue saving page with language that was not exactly `2` or `5` chars long [#1667](https://github.com/getgrav/grav-plugin-admin/issues/1667)
+    * Fixed admin not detecting any existing users when Flex users are being used
+    * Fixed issue with append/prepend not respecting `size:`
+    * Fixed issue with `unset` on file fields [#1427](https://github.com/getgrav/grav/issues/1427), [#1670](https://github.com/getgrav/grav/issues/1670), [#1982](https://github.com/getgrav/grav/issues/1982)
+
 # v1.9.3
 ## 04/22/2019
 
@@ -5,7 +22,7 @@
     * Added a new **YAML Linter** report to the `Tools - Reports` section
 1. [](#improved)
     * Updated package.json scripts to properly use gulp compiler
-
+    
 # v1.9.2
 ## 04/15/2019